Moved remote_resouce to module
