+++ setup.py
@@ -46,7 +46,6 @@ client = set([
 develop = set([
     'futures==3.0.5',
     'requests==2.13.0',
-    'requests_cache',
     'wheel',
 ])
